Homeの設定
・メイン画面の名称をHome screen に変更
・home.dartファイルを作成
