+++ windows/runner/resource.h
@@ -4,13 +4,13 @@
 //
 #define IDI_APP_ICON                    101
 
-// Next default values for new objects
+// bar default values for new objects
 //
 #ifdef APSTUDIO_INVOKED
 #ifndef APSTUDIO_READONLY_SYMBOLS
-#define _APS_NEXT_RESOURCE_VALUE        102
-#define _APS_NEXT_COMMAND_VALUE         40001
-#define _APS_NEXT_CONTROL_VALUE         1001
-#define _APS_NEXT_SYMED_VALUE           101
+#define _APS_bar_RESOURCE_VALUE        102
+#define _APS_bar_COMMAND_VALUE         40001
+#define _APS_bar_CONTROL_VALUE         1001
+#define _APS_bar_SYMED_VALUE           101
 #endif
 #endif
